modules/kafka-producer/requirements.txt
kafka-python==2.0.2
grpcio==1.40.0
grpcio-tools == 1.34.0
protobuf==3.19.5